alter: allow to restrict space format of a non-empty space
Now if a space is not empty, its format can not be restricted by
nullability removal, or by restriction of a field type.

Allow format restriction followed by space format checking.

Do not rebuild index, if it is not empty. A rebuild is not needed
for such change, because types can not be altered on a not empty space
in such way, that index order is changed.

But do rebuild, if an index is not unique, and a primary index parts
have been changed. It changes cmp_def of the secondary index, and
possibly sort order.

Needed for #3008