schema: allow to define multiple types for one space field
On a single field two restrictions exist: tuple format and index part.

The proposal is to allow to use different types for different restrictions,
if these types are compatible. And for tuples validation use the most
strict type.

For example, one field can be number in space format, unsigned in one index,
integer in another index and scalar in a third index. Configuration like can be
allowed now, because all of these types can store unsigned. And unsigned is
used to validate tuples.

Besides, it allows incrementaly DDL. For example, if you want to change a
type of a field, which is used by several indexes, then it is impossible on
a non-empty space, even if a new type is less strict than the old one.

Closes #3008